Added a function, GetLastMessageID, for accessing the last message ID
Added a function, GetLastMessageID, for accessing the last message ID.  It returns whatever message ID was used when the headers were last created, which is also the message ID of the last sent message except in failure cases.  If the client sends multiple messages using the same PHPMailer object, the value returned by GetLastMessageID() will be updated each time.  In issue #57, it was pointed out that a client wanting the message ID could just explicitly set the message ID via $this->MessageID, but that would require the client to understand how to create a good message ID.